requirements.txt
llvmlite==0.47.0
numba==0.65.1
numpy==2.4.4
pillow==12.2.0
